Move input manager opaque types
At last some of these opaque types are used in other modules.
