+++ Cargo.toml
@@ -15,7 +15,7 @@ members = ["components/*", "components/multimedia/image_framework", "components/
 exclude = ["components/multimedia", "scripts"]
 
 [workspace.dependencies]
-ohos-sys-opaque-types = { version = "0.1.5", path = "components/opaque-types" }
+ohos-sys-opaque-types = { version = "0.1.6", path = "components/opaque-types" }
 
 
 [dependencies]
@@ -28,7 +28,7 @@ ohos-deviceinfo-sys = { version = "0.1.2", optional = true, path = "components/d
 ohos-drawing-sys = { path = "components/drawing", version = "0.3.0", optional = true}
 ohos-image-kit-sys = {version = "0.2", optional = true, path = "components/multimedia/image_framework"}
 ohos-ime-sys = { version = "0.2.0", optional = true, path = "components/inputmethod"}
-ohos-input-sys = { version = "0.2.0", optional = true, path  = "components/multimodal-input"}
+ohos-input-sys = { version = "0.3.0", optional = true, path  = "components/multimodal-input"}
 ohos-rawfile-sys = { version = "0.1.0", optional = true, path = "components/rawfile"}
 ohos-sys-opaque-types = { workspace = true}
 ohos-window-sys = { version = "0.1", optional = true, path = "components/window"}
